src/requirements.txt
transformers==4.40.1
pyannote.audio==3.1.1
python-docx==1.1.0
intervaltree==3.1.0

openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd

--index-url https://download.pytorch.org/whl/cu121
torch==2.3.0+cu121
